refactor(control-api): extract access/fetch/delivery/execute from response_actions/mod.rs
