+++ crates/services/control-api/src/routes/response_actions/mod.rs
@@ -43,10 +43,18 @@ terminate_session, kill_switch, quarantine_principal, revoke_grant, revoke_princ
 policy_rule_diff_validation";
 const ACK_STATUS_ALLOWLIST: &str = "acknowledged, rejected, failed, expired, rolled_back";
 
+mod access;
+mod delivery;
 mod dto;
+mod execute;
+mod fetch;
 
+pub(crate) use access::*;
+pub(crate) use delivery::*;
 pub(crate) use dto::*;
 pub use dto::{CreateResponseActionRequest, ResponseTargetInput};
+pub(crate) use execute::*;
+pub(crate) use fetch::*;
 
 pub fn router() -> Router<AppState> {
     Router::new()
@@ -280,20 +288,6 @@ async fn publish_action(
     Ok(Json(
         fetch_action_detail(state, tenant_id, action_id).await?,
     ))
-}
-
-fn ensure_write_access(auth: &AuthenticatedTenant) -> Result<(), ApiError> {
-    if !matches!(auth.role.as_str(), "owner" | "admin") {
-        return Err(ApiError::Forbidden);
-    }
-    Ok(())
-}
-
-fn ensure_api_key_executor(auth: &AuthenticatedTenant) -> Result<(), ApiError> {
-    if auth.role == "viewer" || !auth.is_api_key() {
-        return Err(ApiError::Forbidden);
-    }
-    Ok(())
 }
 
 async fn prepare_create_action(
@@ -2145,412 +2139,6 @@ fn ensure_publishable(current_status: &str, allow_retry: bool) -> Result<(), Api
             "status '{status}' cannot be approved"
         ))),
     }
-}
-
-fn delivery_plan(action: &ResponseActionRecord, tenant_slug: &str) -> DeliveryPlan {
-    let subject_prefix = tenant_subject_prefix(tenant_slug);
-    let ack_deadline = action
-        .require_acknowledgement
-        .then(|| Utc::now() + Duration::minutes(ACK_DEADLINE_MINUTES));
-    let ack_token = action
-        .require_acknowledgement
-        .then(|| Uuid::new_v4().to_string());
-
-    match action.target.kind {
-        ResponseTargetKind::Endpoint
-        | ResponseTargetKind::Runtime
-        | ResponseTargetKind::Session => {
-            let canonical_subject = canonical_response_subject(action, &subject_prefix);
-            let legacy_subject = legacy_posture_subject(action, &subject_prefix);
-
-            DeliveryPlan {
-                target_kind: action.target.kind.as_str().to_string(),
-                target_id: action.target.id.clone(),
-                executor_kind: match action.target.kind {
-                    ResponseTargetKind::Endpoint => "endpoint_agent".to_string(),
-                    ResponseTargetKind::Runtime => "runtime_agent".to_string(),
-                    ResponseTargetKind::Session => "session_api".to_string(),
-                    _ => "endpoint_agent".to_string(),
-                },
-                delivery_subject: Some(canonical_subject.clone()),
-                acknowledgement_deadline: ack_deadline,
-                metadata: json!({
-                    "ack_token": ack_token,
-                    "canonical_subject": canonical_subject,
-                    "compat_mirror_subject": legacy_subject,
-                    "protocol": "response_action_v1",
-                }),
-            }
-        }
-        _ => DeliveryPlan {
-            target_kind: action.target.kind.as_str().to_string(),
-            target_id: action.target.id.clone(),
-            executor_kind: "cloud_only".to_string(),
-            delivery_subject: None,
-            acknowledgement_deadline: ack_deadline,
-            metadata: json!({
-                "ack_token": ack_token,
-                "cloud_only": true,
-                "protocol": "cloud_only",
-            }),
-        },
-    }
-}
-
-fn canonical_response_subject(action: &ResponseActionRecord, subject_prefix: &str) -> String {
-    format!(
-        "{subject_prefix}.response.command.{}.{}",
-        action.target.kind.as_str(),
-        action.target.id
-    )
-}
-
-fn legacy_posture_subject(action: &ResponseActionRecord, subject_prefix: &str) -> Option<String> {
-    if matches!(
-        action.action_type.as_str(),
-        "transition_posture" | "request_policy_reload" | "kill_switch"
-    ) && matches!(action.target.kind, ResponseTargetKind::Endpoint)
-    {
-        Some(format!(
-            "{subject_prefix}.posture.command.{}",
-            action.target.id
-        ))
-    } else {
-        None
-    }
-}
-
-fn build_delivery_payload_bytes(
-    action: &ResponseActionRecord,
-    delivery: &ResponseActionDelivery,
-    signing_enabled: bool,
-    signing_keypair: Option<&hush_core::Keypair>,
-) -> Result<Vec<u8>, ApiError> {
-    let payload = action_transport_payload(action, delivery);
-    build_signed_payload_bytes(payload, signing_enabled, signing_keypair)
-}
-
-fn action_transport_payload(
-    action: &ResponseActionRecord,
-    delivery: &ResponseActionDelivery,
-) -> Value {
-    let ack_token = delivery.metadata.get("ack_token").and_then(Value::as_str);
-
-    let mut payload = action.to_transport_payload();
-    payload["delivery"] = json!({
-        "subject": delivery.delivery_subject,
-        "targetKind": delivery.target_kind,
-        "targetId": delivery.target_id,
-        "ackToken": ack_token,
-    });
-    payload
-}
-
-async fn execute_cloud_only_action(
-    state: &AppState,
-    action: &ResponseActionRecord,
-) -> Result<DeliveryExecution, ApiError> {
-    match (
-        ResponseActionType::from_str(&action.action_type)?,
-        &action.target.kind,
-    ) {
-        (ResponseActionType::RevokeGrant, ResponseTargetKind::Grant) => {
-            let grant_id = Uuid::parse_str(&action.target.id).map_err(|_| {
-                ApiError::BadRequest("grant targets must use a UUID grant id".to_string())
-            })?;
-            let response = delegation_graph_service::revoke_grant(
-                &state.db,
-                action.tenant_id,
-                grant_id,
-                RevokeGrantRequest {
-                    reason: action.reason.clone(),
-                    revoke_descendants: Some(true),
-                    revoked_by: Some(action.requested_by.actor_id.clone()),
-                    response_action_id: Some(action.id.to_string()),
-                    response_action_label: Some(action.action_type.clone()),
-                    response_action_state: Some("acknowledged".to_string()),
-                    response_action_metadata: Some(json!({
-                        "response_action_id": action.id.to_string(),
-                        "action_type": action.action_type.as_str(),
-                    })),
-                },
-            )
-            .await?;
-
-            Ok(DeliveryExecution::Acknowledged {
-                observed_at: Utc::now(),
-                message: Some("grant revoked".to_string()),
-                resulting_state: Some("revoked".to_string()),
-                raw_payload: json!({
-                    "grantId": grant_id,
-                    "revokedGrantIds": response.revoked_grant_ids,
-                    "status": "acknowledged",
-                }),
-            })
-        }
-        (ResponseActionType::QuarantinePrincipal, ResponseTargetKind::Principal) => {
-            execute_principal_lifecycle_action(state, action, "quarantined").await
-        }
-        (ResponseActionType::RevokePrincipal, ResponseTargetKind::Principal) => {
-            execute_principal_lifecycle_action(state, action, "revoked").await
-        }
-        _ => Err(ApiError::BadRequest(format!(
-            "action '{}' does not have an executable control-plane handler",
-            action.action_type
-        ))),
-    }
-}
-
-async fn execute_principal_lifecycle_action(
-    state: &AppState,
-    action: &ResponseActionRecord,
-    lifecycle_state: &str,
-) -> Result<DeliveryExecution, ApiError> {
-    let target = update_principal_lifecycle_target(
-        &state.db,
-        action.tenant_id,
-        &action.target.id,
-        lifecycle_state,
-    )
-    .await?;
-    sync_principal_graph_state(&state.db, action.tenant_id, &target, lifecycle_state).await?;
-    let revoked_grant_ids = delegation_graph_service::revoke_principal_grants(
-        &state.db,
-        action.tenant_id,
-        delegation_graph_service::RevokePrincipalGrantsRequest {
-            principal_id: target.principal_id,
-            principal_stable_ref: target.stable_ref.clone(),
-            reason: action.reason.clone(),
-            revoked_by: Some(action.requested_by.actor_id.clone()),
-            response_action_id: Some(action.id.to_string()),
-            response_action_label: Some(action.action_type.clone()),
-            response_action_state: Some("acknowledged".to_string()),
-            response_action_metadata: Some(json!({
-                "response_action_id": action.id.to_string(),
-                "action_type": action.action_type.as_str(),
-            })),
-        },
-    )
-    .await?;
-
-    Ok(DeliveryExecution::Acknowledged {
-        observed_at: Utc::now(),
-        message: Some(format!("principal transitioned to {lifecycle_state}")),
-        resulting_state: Some(lifecycle_state.to_string()),
-        raw_payload: json!({
-            "principalId": target.principal_id,
-            "stableRef": target.stable_ref,
-            "status": "acknowledged",
-            "lifecycleState": lifecycle_state,
-            "revokedGrantIds": revoked_grant_ids,
-        }),
-    })
-}
-
-async fn update_principal_lifecycle_target(
-    db: &crate::db::PgPool,
-    tenant_id: Uuid,
-    target_id: &str,
-    lifecycle_state: &str,
-) -> Result<PrincipalLifecycleTarget, ApiError> {
-    let principal_id = Uuid::parse_str(target_id.trim()).map_err(|_| {
-        ApiError::Internal("principal response targets must be canonical UUID ids".to_string())
-    })?;
-    let row = sqlx::query::query(
-        r#"UPDATE principals
-           SET lifecycle_state = $3,
-               updated_at = now()
-           WHERE tenant_id = $1
-             AND id = $2
-           RETURNING id, stable_ref"#,
-    )
-    .bind(tenant_id)
-    .bind(principal_id)
-    .bind(lifecycle_state)
-    .fetch_optional(db)
-    .await
-    .map_err(ApiError::Database)?
-    .ok_or(ApiError::NotFound)?;
-
-    Ok(PrincipalLifecycleTarget {
-        principal_id: row.try_get("id").map_err(ApiError::Database)?,
-        stable_ref: row.try_get("stable_ref").map_err(ApiError::Database)?,
-    })
-}
-
-async fn sync_principal_graph_state(
-    db: &crate::db::PgPool,
-    tenant_id: Uuid,
-    target: &PrincipalLifecycleTarget,
-    lifecycle_state: &str,
-) -> Result<(), ApiError> {
-    let node_ids = vec![
-        format!("principal:{}", target.principal_id),
-        format!("principal:{}", target.stable_ref),
-    ];
-    sqlx::query::query(
-        r#"UPDATE delegation_graph_nodes
-           SET state = $3,
-               updated_at = now()
-           WHERE tenant_id = $1
-             AND id = ANY($2)"#,
-    )
-    .bind(tenant_id)
-    .bind(&node_ids)
-    .bind(lifecycle_state)
-    .execute(db)
-    .await
-    .map_err(ApiError::Database)?;
-    Ok(())
-}
-
-fn legacy_posture_command_payload(action: &ResponseActionRecord) -> Result<Value, ApiError> {
-    match action.action_type.as_str() {
-        "transition_posture" => {
-            let posture = transition_posture_value(&action.payload).ok_or_else(|| {
-                ApiError::BadRequest(
-                    "transition_posture actions require payload.toState or payload.posture"
-                        .to_string(),
-                )
-            })?;
-            Ok(json!({
-                "command": "set_posture",
-                "posture": posture,
-            }))
-        }
-        "request_policy_reload" => Ok(json!({
-            "command": "request_policy_reload",
-        })),
-        "kill_switch" => Ok(json!({
-            "command": "kill_switch",
-            "reason": action.reason,
-        })),
-        other => Err(ApiError::BadRequest(format!(
-            "action '{other}' does not support legacy posture transport"
-        ))),
-    }
-}
-
-fn build_signed_payload_bytes(
-    payload: Value,
-    signing_enabled: bool,
-    signing_keypair: Option<&hush_core::Keypair>,
-) -> Result<Vec<u8>, ApiError> {
-    if signing_enabled {
-        let keypair = signing_keypair.ok_or_else(|| {
-            ApiError::Internal("response signing is enabled but keypair is not loaded".to_string())
-        })?;
-        let envelope =
-            spine::build_signed_envelope(keypair, 0, None, payload, spine::now_rfc3339()).map_err(
-                |err| ApiError::Internal(format!("failed to sign response action: {err}")),
-            )?;
-        return serde_json::to_vec(&envelope).map_err(|err| {
-            ApiError::Internal(format!(
-                "failed to serialize signed response action envelope: {err}"
-            ))
-        });
-    }
-
-    Ok(serde_json::to_vec(&payload).unwrap_or_default())
-}
-
-fn transition_posture_value(payload: &Value) -> Option<String> {
-    payload
-        .get("toState")
-        .and_then(Value::as_str)
-        .or_else(|| payload.get("to_state").and_then(Value::as_str))
-        .or_else(|| payload.get("posture").and_then(Value::as_str))
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(ToOwned::to_owned)
-}
-
-fn scrub_delivery_metadata(mut metadata: Value) -> Value {
-    if let Some(object) = metadata.as_object_mut() {
-        object.remove("ack_token");
-    }
-    metadata
-}
-
-async fn fetch_action(
-    state: &AppState,
-    tenant_id: Uuid,
-    action_id: Uuid,
-) -> Result<ResponseActionRecord, ApiError> {
-    let row = sqlx::query::query("SELECT * FROM response_actions WHERE tenant_id = $1 AND id = $2")
-        .bind(tenant_id)
-        .bind(action_id)
-        .fetch_optional(&state.db)
-        .await
-        .map_err(ApiError::Database)?
-        .ok_or(ApiError::NotFound)?;
-    ResponseActionRecord::from_row(row).map_err(ApiError::Database)
-}
-
-async fn fetch_deliveries(
-    state: &AppState,
-    tenant_id: Uuid,
-    action_id: Uuid,
-) -> Result<Vec<ResponseActionDelivery>, ApiError> {
-    let rows = sqlx::query::query(
-        r#"SELECT *
-           FROM response_action_deliveries
-           WHERE tenant_id = $1 AND action_id = $2
-           ORDER BY created_at DESC, id DESC"#,
-    )
-    .bind(tenant_id)
-    .bind(action_id)
-    .fetch_all(&state.db)
-    .await
-    .map_err(ApiError::Database)?;
-
-    rows.into_iter()
-        .map(ResponseActionDelivery::from_row)
-        .map(|delivery| {
-            let mut delivery = delivery?;
-            delivery.metadata = scrub_delivery_metadata(delivery.metadata);
-            Ok(delivery)
-        })
-        .collect::<Result<Vec<_>, _>>()
-        .map_err(ApiError::Database)
-}
-
-async fn fetch_acks(
-    state: &AppState,
-    tenant_id: Uuid,
-    action_id: Uuid,
-) -> Result<Vec<ResponseActionAckRecord>, ApiError> {
-    let rows = sqlx::query::query(
-        r#"SELECT *
-           FROM response_action_acks
-           WHERE tenant_id = $1 AND action_id = $2
-           ORDER BY observed_at DESC, id DESC"#,
-    )
-    .bind(tenant_id)
-    .bind(action_id)
-    .fetch_all(&state.db)
-    .await
-    .map_err(ApiError::Database)?;
-
-    rows.into_iter()
-        .map(ResponseActionAckRecord::from_row)
-        .collect::<Result<Vec<_>, _>>()
-        .map_err(ApiError::Database)
-}
-
-async fn fetch_action_detail(
-    state: &AppState,
-    tenant_id: Uuid,
-    action_id: Uuid,
-) -> Result<ResponseActionDetail, ApiError> {
-    let action = fetch_action(state, tenant_id, action_id).await?;
-    let deliveries = fetch_deliveries(state, tenant_id, action_id).await?;
-    let acknowledgements = fetch_acks(state, tenant_id, action_id).await?;
-    Ok(ResponseActionDetail {
-        action,
-        deliveries,
-        acknowledgements,
-    })
 }
 
 async fn link_action_to_source_detection(
